Conflicts:
	main.cc

+++ optimizedKernel.h
@@ -1,6 +1,6 @@
 #ifndef OPTIMIZEDKERNEL_H
 #define OPTIMIZEDKERNEL_H
 
-int optimizedKernelSetup(int, int, bool = false);
+float optimizedKernelSetup(int, int, bool = false);
 
 #endif
